# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Connecting Neighbors - Hyperlocal Community App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  replitId: string;
  name: string;
  username: string;
  bio?: string | null;
  apartment?: string | null;
  avatarUrl?: string | null;
  phone?: string | null;
  isVerified: boolean;
  joinedAt: string;
}

export interface UpdateUserBody {
  name?: string;
  bio?: string;
  apartment?: string;
  phone?: string;
  avatarUrl?: string;
}

export type PostCategory = (typeof PostCategory)[keyof typeof PostCategory];

export const PostCategory = {
  general: "general",
  announcement: "announcement",
  helpNeeded: "helpNeeded",
  lostFound: "lostFound",
  recommendation: "recommendation",
  safety: "safety",
} as const;

export interface Post {
  id: number;
  authorId: number;
  author: User;
  title?: string | null;
  content: string;
  category: PostCategory;
  imageUrl?: string | null;
  likesCount: number;
  commentsCount: number;
  isLikedByMe: boolean;
  createdAt: string;
}

export type CreatePostBodyCategory =
  (typeof CreatePostBodyCategory)[keyof typeof CreatePostBodyCategory];

export const CreatePostBodyCategory = {
  general: "general",
  announcement: "announcement",
  helpNeeded: "helpNeeded",
  lostFound: "lostFound",
  recommendation: "recommendation",
  safety: "safety",
} as const;

export interface CreatePostBody {
  title?: string;
  content: string;
  category: CreatePostBodyCategory;
  imageUrl?: string;
}

export interface LikeResponse {
  liked: boolean;
  likesCount: number;
}

export type ListingType = (typeof ListingType)[keyof typeof ListingType];

export const ListingType = {
  sell: "sell",
  free: "free",
  rent: "rent",
} as const;

export type ListingCategory =
  (typeof ListingCategory)[keyof typeof ListingCategory];

export const ListingCategory = {
  electronics: "electronics",
  furniture: "furniture",
  clothing: "clothing",
  books: "books",
  groceries: "groceries",
  appliances: "appliances",
  other: "other",
} as const;

export interface Listing {
  id: number;
  sellerId: number;
  seller: User;
  title: string;
  description: string;
  price?: number | null;
  type: ListingType;
  category: ListingCategory;
  imageUrl?: string | null;
  isAvailable: boolean;
  createdAt: string;
}

export type CreateListingBodyType =
  (typeof CreateListingBodyType)[keyof typeof CreateListingBodyType];

export const CreateListingBodyType = {
  sell: "sell",
  free: "free",
  rent: "rent",
} as const;

export type CreateListingBodyCategory =
  (typeof CreateListingBodyCategory)[keyof typeof CreateListingBodyCategory];

export const CreateListingBodyCategory = {
  electronics: "electronics",
  furniture: "furniture",
  clothing: "clothing",
  books: "books",
  groceries: "groceries",
  appliances: "appliances",
  other: "other",
} as const;

export interface CreateListingBody {
  title: string;
  description: string;
  price?: number;
  type: CreateListingBodyType;
  category: CreateListingBodyCategory;
  imageUrl?: string;
}

export interface UpdateListingBody {
  title?: string;
  description?: string;
  price?: number;
  isAvailable?: boolean;
}

export interface Event {
  id: number;
  organizerId: number;
  organizer: User;
  title: string;
  description: string;
  location: string;
  startsAt: string;
  endsAt?: string | null;
  rsvpCount: number;
  isRsvpedByMe: boolean;
  createdAt: string;
}

export interface CreateEventBody {
  title: string;
  description: string;
  location: string;
  startsAt: string;
  endsAt?: string;
}

export interface RsvpResponse {
  rsvped: boolean;
  rsvpCount: number;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  emergency: "emergency",
} as const;

export interface Alert {
  id: number;
  reporterId: number;
  reporter: User;
  title: string;
  description: string;
  severity: AlertSeverity;
  isResolved: boolean;
  createdAt: string;
}

export type CreateAlertBodySeverity =
  (typeof CreateAlertBodySeverity)[keyof typeof CreateAlertBodySeverity];

export const CreateAlertBodySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  emergency: "emergency",
} as const;

export interface CreateAlertBody {
  title: string;
  description: string;
  severity: CreateAlertBodySeverity;
}

export type ResourceType = (typeof ResourceType)[keyof typeof ResourceType];

export const ResourceType = {
  ride: "ride",
  item: "item",
  service: "service",
  childcare: "childcare",
} as const;

export interface Resource {
  id: number;
  offererId: number;
  offerer: User;
  title: string;
  description: string;
  type: ResourceType;
  isAvailable: boolean;
  createdAt: string;
}

export type CreateResourceBodyType =
  (typeof CreateResourceBodyType)[keyof typeof CreateResourceBodyType];

export const CreateResourceBodyType = {
  ride: "ride",
  item: "item",
  service: "service",
  childcare: "childcare",
} as const;

export interface CreateResourceBody {
  title: string;
  description: string;
  type: CreateResourceBodyType;
}

export interface CommunityStats {
  totalMembers: number;
  totalPosts: number;
  activeListings: number;
  upcomingEvents: number;
  activeAlerts: number;
  availableResources: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  post: "post",
  listing: "listing",
  event: "event",
  alert: "alert",
  resource: "resource",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  actorName: string;
  actorAvatar?: string | null;
  createdAt: string;
}

export type ListPostsParams = {
  category?: string;
  limit?: number;
  offset?: number;
};

export type ListListingsParams = {
  category?: string;
  type?: ListListingsType;
};

export type ListListingsType =
  (typeof ListListingsType)[keyof typeof ListListingsType];

export const ListListingsType = {
  sell: "sell",
  free: "free",
  rent: "rent",
} as const;

export type ListResourcesParams = {
  type?: ListResourcesType;
};

export type ListResourcesType =
  (typeof ListResourcesType)[keyof typeof ListResourcesType];

export const ListResourcesType = {
  ride: "ride",
  item: "item",
  service: "service",
  childcare: "childcare",
} as const;