# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Connecting Neighbors - Hyperlocal Community App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List community feed posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /marketplace:
    get:
      operationId: listListings
      tags: [marketplace]
      summary: List marketplace items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [sell, free, rent]
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [marketplace]
      summary: Create a marketplace listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /marketplace/{id}:
    get:
      operationId: getListing
      tags: [marketplace]
      summary: Get a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    put:
      operationId: updateListing
      tags: [marketplace]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingBody"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [marketplace]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List community events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a community event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RSVP toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RsvpResponse"
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List shared resources (rides, items, services)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [ride, item, service, childcare]
      responses:
        "200":
          description: Resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Offer a shared resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Remove a shared resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        replitId:
          type: string
        name:
          type: string
        username:
          type: string
        bio:
          type: string
          nullable: true
        apartment:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        isVerified:
          type: boolean
        joinedAt:
          type: string
          format: date-time
      required: [id, replitId, name, username, isVerified, joinedAt]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        title:
          type: string
          nullable: true
        content:
          type: string
        category:
          type: string
          enum: [general, announcement, helpNeeded, lostFound, recommendation, safety]
        imageUrl:
          type: string
          nullable: true
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLikedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, authorId, author, content, category, likesCount, commentsCount, isLikedByMe, createdAt]
    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [general, announcement, helpNeeded, lostFound, recommendation, safety]
        imageUrl:
          type: string
      required: [content, category]
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required: [liked, likesCount]
    Listing:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        price:
          type: number
          nullable: true
        type:
          type: string
          enum: [sell, free, rent]
        category:
          type: string
          enum: [electronics, furniture, clothing, books, groceries, appliances, other]
        imageUrl:
          type: string
          nullable: true
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, sellerId, seller, title, description, type, category, isAvailable, createdAt]
    CreateListingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        type:
          type: string
          enum: [sell, free, rent]
        category:
          type: string
          enum: [electronics, furniture, clothing, books, groceries, appliances, other]
        imageUrl:
          type: string
      required: [title, description, type, category]
    UpdateListingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
    Event:
      type: object
      properties:
        id:
          type: integer
        organizerId:
          type: integer
        organizer:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
          nullable: true
        rsvpCount:
          type: integer
        isRsvpedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, organizerId, organizer, title, description, location, startsAt, rsvpCount, isRsvpedByMe, createdAt]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
      required: [title, description, location, startsAt]
    RsvpResponse:
      type: object
      properties:
        rsvped:
          type: boolean
        rsvpCount:
          type: integer
      required: [rsvped, rsvpCount]
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        offererId:
          type: integer
        offerer:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [ride, item, service, childcare]
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, offererId, offerer, title, description, type, isAvailable, createdAt]
    CreateResourceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [ride, item, service, childcare]
      required: [title, description, type]
...